requirements.txt
fuzzywuzzy==0.15.0
pymssql==2.1.3
